User groups are not pre-selected when opening the permissions modal from the embedded display view
The issue was due to a difference with the data bound when the component is
mounted.
When it is mount from one of the dropdown menu in the general view there is
no data for the item to update and the item is updated right after. When the
modal is mount from the embedded display view, the actual item to update is
already present when the component is mounted.

Part of story #12457: New document modal: permissions

Change-Id: Ibc4a7cea4e1a12c6c260683e302517f74508ecf4