Define by project if the burnup will be displayed in count or effort mode
This is part of story #13735 have burn-up based on item count

If the feature flag (use_burnup_count_elements) is set to 1
Then a new option named charts configuration is displayed in the global
agiledashboard configuration to enable the burnup count mode.

If this option is activated, then all the project's burnup will be
displayed in count mode.

This option is also inherited at project creation (not yet in XML).

Change-Id: If1201f645a4046cf83d6826dc70b602591c9228d