Avoid display of internal key to end user
When I update the label or description of a service, I should not see
the internal key.

If I don't touch the label or description, then in database the internal
key is kept. However if I change the label or description then in the
database the new version is saved.

Part of story #13567 decide that custom services opens in a new tab/window by
default

Change-Id: Ia5cf8a1f9bdcc569086804eecacd0397d31e526b