Select icon and open in new tab in Add modal
part of story #13566 associate a custom icon to external services
part of story #13567 decide that custom services opens in a new
tab/window by default

How to test:
- As project admin, you can add a new service. The icon selector will let
you choose an icon and there is a switch to "open in a new tab" but those
values are not saved yet (they are ignored). There is a preview of the
label and icon.
- As site admin, you can add a new service and can also set it to
available for new projects.
- As site admin, only in the Default Site Template, you can choose a
short name. However, when you add it, the service is invisible in the UI
of all projects [0].
- You can no longer choose to add a new service that opens in an iframe
(the input is removed).
- No functional change when Editing services.

[0]: see request #13712, that behaviour is the same on master.

Change-Id: I13f5f9678524cb160e77f5115458ff38f784d6a0