I can update Admin service
I should be able to change the rank of the Admin service.

The issue was coming from the fact that is_used was not sent by the front
end (because this service cannot be unused). Therefore backend should
force is_used to true.

Part of story #13567: decide that custom services opens in a new tab/window by default

Change-Id: Id21f631bf5b94c2cd918b6fe4427914e36048444