Remove help block for link
We do not want to promote too much the usage of substitution variables in
service links. Those variables are only useable in template and should
not be displayed to regular project admins.

Since it is a recent addition, no regression is expected from this
removal.

Part of story #13567 decide that custom services opens in a new tab/window by
default

Change-Id: I9842ce40d7c924a0fc80bd9f223703e6701f16cb