Inherit icon and "new tab" in new project
part of story #13566 associate a custom icon to external services
part of story #13567 decide that custom services opens in a new
tab/window by default

How to test:
- Create a custom service in a "source" project with an icon and that
opens in a new tab.
- Create a new "destination" project from the "source" project.
- "destination" should have the same custom service with the same icon
and it should open in a new tab.

Note: Given that 'server_id' and 'location' are marked as "To be
deleted", they are no longer copied when inheriting a project.

Change-Id: Ic7ad4c70295e3e2ea16db4948842c92871a8172d