Updating the current folder should not be rewritten immediatly by a folder content reload
When updating the the current folder, only a part of the object was replaced
in the "current folder ascendant hierarchy" and not the whole reference. As a
result when you update the current folder and ask for refresh of the document
in it, the "current folder ascendant hierarchy" consider the current folder is
outdated because the object references does not match and replace it with what
can be the reference it currently holds (which is a partially updated information).
To avoid that, when replacing the current folder, the reference in the ascendant
hierarchy should be replaced instead of partially updating the object.

It's possible to see the issue by updating the permissions or metadata of the
current folder somewhere in the hierarchy and opening again the modal after the
update: outdated information will be shown. Note that the issue is not present when
working on the root folder because in this case the ascendant hierarchy is empty.

Part of story #12457: New document modal: permissions

Change-Id: I7a245b3b3cfb9101faa0dcca5ab0a28b71eca7d7