Force title to be a string
Fix request #13772: null is displayed as title of a node in graph of dependencies

In some situations, the title returned by the testmanagement_nodes can
be null instead of empty string.

This null value is displayed as is in the graph of dependencies modal :(

You can test this for example with the create_test_env plugin.

Change-Id: Ic88f88a893a3d8b9fb1ec662eba2d065eb44f7a2