Make the static analysis tools work on the whole codebase when a massive change is processed
With too much changed files the test pipeline crashes because
the list of files given as an argument to the commands is too long.
We could fix that but it's not worth it, the analysises are going to be
slow anyway so let's just scan everything.

Needed for request #13780: Enforce the position of arguments and braces of
methods in the PHP codebase

Change-Id: I2671a6b24eb1f71d96be71b4c3a77ac1845b8d56