Fix 6 issues detected by Psalm already existing in the codebase
Those 6 issues are set in the Psalm's baseline and, as such, not reported
in usual run. request #13780: Enforce the position of arguments and braces
of methods comes with a massive change and those 3 errors are no more detected
known issues. This contribution fixes that before applying the style fix everywhere.

Change-Id: I0ee7d59fd5fad156261b8cb84b96cf6b18058914