Move the release widget plugin to Jest 🃏
The test files have been moved from *.spec.ts to *.test.ts
to ease the configuration of ESLint and to make it easy to
spot tests that have not yet been migrated from Karma/Jasmine.

This contribution does yet not configure properly the snapshotting
of Vue components. It will come with another commit and an example.

Code coverage is also not collected in the CI pipeline.

Part of request #13789: Move the release widget plugin to Jest

Change-Id: Iafe68fb8d26ad0790d8d07c269cd142e361b65b0