request #13845 Replace easygettext by our own gettext extractor
How to test:
- make generate-po should not result in any changes in .po files.

Notes:
- A translation string was never extracted from document app, I took the
liberty of translating it to french.
- It extracts from [v-translate], <translate> and title="$gettext()"
- Some translations contain HTML tags. vue-gettext allows it inside
[v-translate] directives, but the extraction of such strings requires us
to parse HTML _in addition_ to parsing Vue.js' AST. Given that there is
a workable solution around that problem (using $gettext and
$gettextInterpolate + v-dompurify-html), supporting this use-case is not
worth the effort. The upside is we can just drop HTML-parsing
dependencies such as cheerio and pug.

Change-Id: I1dca7e25fbb366dec94e148ba804ac22e4286fe3