Setup snapshot testing of Vue components
A limit to the size of a snapshot is enforced via ESLint. A snapshot is
supposed to be reviewed, it's not really possible with large snapshots
with a lot of changes with each contributions. Incidently, it also encourages
the developer to work with smaller components instead of one giant blob.

The snapshots are not inspected by the 'no trailing whitespace' pre-commit hook,
it is generated code and whitespaces are important in HTML code.

Part of request #13789: Move the release widget plugin to Jest

Change-Id: I4e5bc18ea54417d9d0223ca10d435397b7aa39be