request #13793: Add kanban entry in user history
Every time I visit a kanban, the user history should be updated.

When user clear the history, or when the kanban is deleted, the
corresponding entries are removed from the table.

Note: only regular view updates the history. If I visit a dashboard with
a kanban widget, then the history is not updated.

Change-Id: I1e0722e774f2b14126f6092e91d79145e6b3eb88