User with REST read only special permission can access all the content through the API
This is part story #13669 permission delegation for a read-only site admin (REST)

If you add directly in the db the new permission delegation (id: 9) for
a given forge user group, then all the members must have access to all
the content through the web UI.

Dedicated checks for each API endpoints will be done.

This is not yet possible to add through the siteamdin web UI this new
permission.

Change-Id: Iefc3dd2b5d1d56785ed7c89dfccc563652d24281