Initiate FrontRouter in Taskboard
This changeset introduces a new route (/taskboard/:project-name/:id)
that is only accessible by typing the url in the browser.

Every milestone accessible to the current user that has a cardwall can
be reached through this url (:id is the id of the milestone). In other
cases: 404.

Part of story #13729: have taskboard empty state

Change-Id: I938cf019bfd0f66478b663a9b84133c5900046b7