Introduce gettext in Taskboard
The under construction message is now localized.

There is an helper to instantiate GettextPlugin, it loads translations
dynamically and asynchronously.

This helper is located in the source of taskboard, but can easily be
moved in a more central place so that it can be used by other
Typescript-Vue app.

My knowledge of Jest + Typescript + Vue is not good enough to be able to
produce unit tests for this helper. I suggest to add them later, for
example when we will move it.

Part of story #13729: have taskboard empty state

Change-Id: I1392c1db7e9aeedc0c824851286a12a403f7cd2b