The new UI should be the UI by default
This is a part of story #13208 create new version of empty document

Before the the old UI was the default UI if the project used custom required metadata.
Now the new docman support the custom metadata feature.

Notice : the user preference for the new UI is not impacted by this patch

How to test:
 - Go to a project with required metadata
=> you should be redirected to the old UI

 - Apply the patch
 - Go to the same project
=> you should be redirected to the new UI

Change-Id: Ifa991ce09c5b1ba0f311cc8454850ec825fe0d81