Mutualize dev/prod configuration of Webpack and make sure IDEs can process TS in test files
Webpack configuration is updated to re-use the same dev and prod configurations
than the other packages that have recently been upgraded. This let us manage the
global set of settings used in these two modes in centralized config file.
Furthermore the configuration of Typescript compiler is updated to drop the exclusion
of the test files. The performance gain is not not worth the hassle of dealing with
two tsconfig to keep the linting working.

Part of story #13729: have taskboard empty state

Change-Id: I5033863b66506a2a0c8bb30fe43d7fc10ddc92af