Add Typescript type definitions for tlp
The declaration for the tlp datePicker are quite permissive.
We would need the get the types from the library itself to properly type it
but this requires to update to another major version. It will be done independently.

select2 has been updated (minor upgrade) to limit the chance to have a mismatch between
the types and the actual implementation.

Part of story #13730: display columns

Change-Id: Idc1d3d6479712d0b8fcb9487d0c02ef8b0808fb1