Better naming of overview.js asset
This asset is the result of the build of scrum-header.js. Naming it
overview.js is wrong since it is not only used in overview pane. It is
used also in pv2 or in taskboard.

Using overview.js is misleading as someone will think that we are
loading unnecessary scripts in taskboard view whereas it was not the
case.

Part of story #13730: display columns

Change-Id: Ia4decfcd4986e390dc67ee5eddd28357f781565c