Run Jest tests sequentially in the CI test pipeline
By default Jest tries to run tests in parallel which is,
most of the time, great. On heavily loaded environnements such
as CI servers it is less interesting and even slow down the execution
because it increases the number of context switches between the processes.
In this context, Jest is not the only runner and other tools are
already running in parallel.
Note this is something known and that is even documented in the Jest
documentation [0].

[0] https://jestjs.io/docs/en/troubleshooting#tests-are-extremely-slow-on-docker-and-or-continuous-integration-ci-server

Part of request #13806: Use Jest as the default test framework for JS unit tests

Change-Id: I8fa5d57a279c93c55365891ab8975716a8082ee9