Remove the bundled CodeMirror code from tlp documentation
This is a follow up to c96548ad2c1dd0bac36cf412ebb29e1eb9cd29d8.

CodeMirror is massive and it takes time for no good reason when collecting
coverage information. It can also be easily retrieved via npm so there
is no need to keep the code inside the repository.

Part of request #13806: Use Jest as the default test framework for JS unit tests

Change-Id: I584d7b2d22fc0176016aad17158bc6a55af86409