Use Composer with a decent version of PHP
While Composer is capable of running with very old version
of PHP, it means Composer is forced to use a series of tricks
for doing so. This is not really great and can lead to issues.

Also the ca-certificates package should be up to date so there is
no TLS verification issue.

Also took the opportunity to bump npm version to same MAJOR.MINOR
used by the main Tuleap build.

enterprise #65117: Tuleap Enterprise build crash on myTuleap Contact Support plugin build

Change-Id: Iaa7b59a7ee6fc61ea37db0c2442b9d115045f066