Include SVG Illustrations in an external, cached file
Part of story #13640: See minimal card display

All inline SVG in Vue components is replaced by <use> tags with [href].
All SVG illustrations are now stored in a "illustrations.svg" file. Its
name is hashed, so it can be cached by browsers and no longer needs to
be loaded in the Vue bundle.
It uses CSS variables (Custom Properties) to style the individual SVG
parts. This is not supported by IE11 [0] but Taskboard does not support
IE11 anyway.

[0]: https://caniuse.com/#feat=css-variables

Change-Id: I2b93e65286c29a95e30631807759a6fcc5ac6eb4