Cache tracker structures to increase the speed of REST testsuite
This is part story #13669 permission delegation for a read-only site admin (REST)

When we need to work on a tracker representation to retrieve a given
field for instance, there is no need to do each time a GET request on
trackers/:id.

All the tracker strucutres are cached during the first setup.

The calls to GET trackers/:id are obviously kept into TrackerTests.

Change-Id: I7e88b9bcfe0665a3533de457dce30647e1dc9804