Move vue-gettext-init into Tuleap core
This contribution changes the way Typescript apps
are built in Tuleap. Until now, the typechecking and the
transpilation was entirely managed through ts-loader, we
now split that into two distincts operations.

This is needed because otherwise we end up doing the typechecking
on a content that is already partilly processed. For example, that
means that an like VueConstructor<V extends Vue> is resolved as
VueConstructor<import('plugins/smthing/vue')> and the code we import
from outside our app will be resolved in its own context so for example
VueConstructor<import('core/smthing/vue')>. The type variable being different,
the two interfaces become two distinct types and the usage of the code imported
from outside our app becomes impossible. By splitting the transpilation and the
typechecking we can use the best of both world. The transpilation can take advantages
of the full Webpack pipeline and the typechecking is not bothered by half-processed
source code (it's also closer to what IDEs display to developers).
This approach is also interesting and we probably would have ended up using it
anyway because it helps improving build performance: the typechecking and the
transpilation can be done in parallel.

The typechecking is limited to the context of the app we are building for two
reasons (signatures of imported functions/methods are of course still used):
 * type conflicts in external dependencies: typechecking imported code means we
need to have the two contexts (the one of the actual app and the one from where
the code is imported) working together which can be tricky. For example, the type
definitions of Node and RequireJS conflict.
 * performances: if a composant is used n times in different apps we would
typecheck it n times. It is a waste of resources.

In the present case, the code of vue-gettext-init is still typechecked when the
tests are executed but this choice in a situation where we have untested code
would result in a not typechecked and not tested code. With our current standards
it is probably fine, code should not be written without tests. Also it is likely
that in a close future we are going to be forced to transpile (and consequently
we could typecheck it as well) our Typescript code so it can be used in our
existing JS apps.

vue-gettext has been updated in src/www/scripts/ so the type definition matches
the one used by the taskboard and release widget applications.

Part of story #13640 See minimal card display

Change-Id: I54aad806ff093432521e828c43d585f6b46f063e