request #13864: Update to DOMPurify 2.0.0
This version is tagged as major due to a breaking change
in the API.
This upgrade is forced by an unexpected change in Chrome 77 [0]
(see implications for DOMPurify in [1]). Even if this is not
supposed to impact Tuleap, it can be hard to troubleshoot in
production since this is caused by a specific version of one
browser. It is best to upgrade now to avoid the potential pain.

[0] https://bugs.chromium.org/p/chromium/issues/detail?id=1002685
[1] https://github.com/cure53/DOMPurify/issues/361

Change-Id: Ibc852c866d23824964ee8c0299ffb8457bf07b1b