Improve tlp.recursiveGet efficiency
Part of story #13640 See minimal card display

How to test:
- recursiveGet will now do a first call, get the total number of
elements and then query all "batches" in parallel.
- It should still return the elements at the end of all requests in the
right order.

Note: "getCollectionCallback" is no longer guaranteed to be called with
elements in the "right order" (since queries are parallel). This should
not be a problem however if the app sorts elements itself when relying
on it to do progressive loading.

Change-Id: I8c093c0b99d7c60627c7e4dde045f340bbdc8da6