request #13866: Stop recording recent visits of the "anonymous user"
The information about recent visits is only used to display the user's history
which is not available when you are not authenticated.
Storing it and maintaining the recent visits list of the user puts an unecessary
write load on the DB. This can be a problem for instances publicly exposed on the
Internet because they can scrapped in parallel by bots. The queries used to clean
up the old entries requires a write lock and can fail under a heavy load resulting
in fatal errors.

Change-Id: Iae605c82c96461886fd6b29cb4b8e0a4e57a5a41