Replace <table> by <div> soup
When there is a lot of columns, then the card should stay with a minimal
width, and there should be an horizontal scroll (like for kanban).

There should not be any regression in the user interface:
- The columns should still take the whole viewport, even when there is only
few swimlanes (or during loading skeletons).
- The header columns should be centered
- The popover for wrong color should be properly displayed

Part of story #13641: See solo cards in their column

Change-Id: Id286112352a1d646317f137f43ac53cc5f81a678