Hide colorpicker text field
As of today, the color of the columns in agiledashboard configuration is
a color picker + a text field. The text field allows the user to enter
an invalid color (either outside the scope of the old palette, or an
invalid color name for the new palette). This text field should be
hidden.

request #13880: We should not let users update manually the color of columns

Change-Id: I8c989fe563f89214c3e9b25027d0867cb9dc79f5