request #13893: Worker can not reconnect to a password protected Redis when the socket timeout is reached
To reproduce the issue you need to:
 * use async notifications
 * a password protected Redis instance
 * a worker doing nothing for more than the time defined in default_socket_timeout

When the socket expires Tuleap will try to reconnect but due to the way
Redis::isConnected() works, the client will still be seen as connected
and Tuleap will recreate the connection from scratch (and therefore will not
authenticate again). This contribution makes sure the Redis client is recreated
when we detect a timeout.

Change-Id: I55f6e652e1fb7dbef507b1dae7d48fe939b36473