Make possible to test at the unit level the pipeline rendering metrics for Prometheus
To move to the enalean/prometheus-client library some logic will need
to be added to clear data collected with the current version of the
library.
This contribution splits the existing code and covers it with tests
so the existing changes can also be tested.

No functional change.

Part of request #13894: Move to enalean/prometheus-client

Change-Id: I9826152ed4ba532dc27bc4ca07b75fd4a074d9ca