request #13901: OpenGraph content should not display HTML tag soup
If your artifact description contains HTML content (like links), then
tags should not be present in the corresponding og: meta tag.

This changes the default behavior of Tracker_Artifact::getDescription().
However, there is only 3 usages:
- In cross tracker search CSV export
- In milestone REST represantation (which seems to never be used)
- In OpenGraph

All those usages do not justify the need to have HTML tags in the
description of the artifact, so we can safely prune them.

Change-Id: I6b0fa4b0f4a36c881d92a538962847a1e40c45bc