Add prometheus histogram for system events and queue events
2 new buckets are created, one for each event type (as the timings will differs).
As of today queue events are likely to take less time than system events hence
the more "aggressive" buckets.

Another major difference between the 2 buckets is that system_events gather the
processing time (time elapsed between the beginning and the end of processing)
while the queue bucket gather the whole lead time (from enqueue to end of processing).

The diffrence is due to the "wait randomization" introduced in system_events that
delay the start of processing to avoid contentions on clock time due to old cron.
In this situation, measuring the lead time on system event would actually mean
measuring the randomization. It's useless because people already knows that the
random number is 4.

Part of request #13891 Gather more metrics about Tuleap backend

Change-Id: I6b169d7f742a2adb3e6190ce04d50791e65477de