Introduce DB tests with phpunit
Current DB integration tests rely on TuleapDbTestCase that is SimpleTest
based. The integration is crappy (auto init of the DB, etc) and SimpleTest
is, well, no longer maintained.

It's time to move on!

This is the first contribution that setup the whole stuff and migrate some
tests.

ProjectImportTest was trashed because the test was actually broken since
ages (it was expecting an Exception but the exception was raised for another
purpose than the one the would have corresponded to the test and there is
obvious way to find by what it should be replaced).

request #13919 Convert simpletest db integration tests to phpunit

Change-Id: I8f12e134affc57b6c8e39de31807d5502278059f