Refactor artifact status retrieval
$artifact->getStatus() is returning the label of the value. This cannot
be used when we need to work with the full object. Therefore this patch
extract the logic to retrieve the status value, but we still return the
lable.

The new object will be used by taskboard to be able to return the
current status of cards, in a dedicated commit.

No functional changes.

Part of story #13641: See solo cards in their column

Change-Id: I39205fff7f50b7607d86f19c6736e3f45d67e1da