Prepare field for no mapping message
Given a top level card
When this card has a status that does not map taskboard columns
And this card has no children
Then a message is displayed so that the user knows that something might
be wrong.

This is a preliminary step so that we can set up styles and such.
Another changeset will display the current status value once the API
REST to have the information will be done.

Part of story #13641: See solo cards in their column

Change-Id: I1708fba588aad93c746abba680bb14fc20d02fb3