Display solo cards in their own column
If a card has children, then it is displayed in the unnamed column,
and there are children skeletons.
If a card has no children, then it is displayed in the column matching
its status (as per the corresponding planning configuration).
If the card has no children, and its status does not map to a column,
then it is displayed in the unnamed column with a no-mapping-message.

Part of story #13641: See solo cards in their column

Change-Id: Ib3ae301e88c9b256f5f498c199bbffa09838d0ff