Handle user-defined mappings on custom fields (not status)
Part of story #13641: See solo cards in their column

How to test:
- Given the milestone planning lets you add "User stories" to
"Releases", remove the Status semantic from "User stories". Configure
the cardwall mapping so that "Status" field is used (or another list
field)
- When you browse the taskboard, the User stories should be in the right
column (and no longer "without status").
- Cards from trackers using a multi-selectbox as "cardwall mapped field"
should be in the right column. When multiple values are selected, for
example "Done, Review" then the first one should be used for the
mapping: "Done".
- No change for trackers using the automatic "duck-typing" Status
semantic mapping. They are still in the right column.

Change-Id: I670fe5fff9e279771ca9e51f3f392fea1eaab986