Display project banner on BurningParrot pages
The message is restricted to some minimal formatting
to be sure the look & feel of the page is not impacted.

The message is always displayed completely and users can not
hide it yet. There is also not yet a megaphone in the header.

To tes the contribution, you can set the message via the REST
route PUT /projects/:project_id/banner.

Part of story #12264: display a custom banner inside project pages

Change-Id: I4d1e408a5304ed4f1003481007accd8267581043