Expand/collapse swimlanes
Every swimlane can be collapsed.

When a swimlane is collapsed, only the xref and the title of the parent
card is displayed (with border color).

The swimlane can then be expanded.

The expand/collapsed is not saved.

The swimlanes that are closed are not automatically collapsed, this will
come in a dedicated changeset.

Known issue that will be managed in a dedicated commit:
- collapsed swimlane is not sticky (it should not horizontally scroll
when there are lot of columns)

Part of story #13957: Expand/collapse swimlane

Change-Id: Ia1777150567cc643a14a494acd4317982bf75f1d