Show skeleton for additional batches of children
When the swimlanes are loaded, there is a global skeleton.

When a swimlane is displayed, then parent card is displayed and there are
skeletons for its children.

When some children are displayed in a column, and the swimlane is still
loading children, then one skeleton is still displayed.

When all children are loaded, then there isn't anymore skeletons in the
swimlane.

Part of story #13627: display children of cards

Change-Id: If148d2901c67042ae51abb0a5860ab171f76876b