Angular/Vue artifact modal: adapt computed field at artifact creation
This is part of story #13569 define a default value for computed fields

If a default value has been set for a computed field, then this field
is open in a "manual value" mode with the input filled with the default
value at artifact creation.

Default value behaviour for computed field is the same as the other
numeric fields (for instance, without submit rights for the user the
default value is correctly saved).

REST and CSV import checks will be done in dedicated commits.

Change-Id: I35358a42fa66c5f9d40ad4b6bcc518fcdeb2e74b