Computed default value is used in artifact creation through REST
This is part of story #13569 define a default value for computed fields

When creating an artifact with REST endoint POST artifacts/
If the computed field has a default value and it's not provided in the body of
the request, then the default value is used.

The same behaviour must occurs if the user does not have the right to
submit this computed field.

Change-Id: I6cee1d0a75d2bc3a81af9ea356b9827acd6cd0b0