User can decide to hide the project banner
While it's not yet possible to close or open the project
banner directly from the web UI, this contribution takes into
account that a user might have closed the banner to stop
seeing it on every pages of the project.

If the project message is updated by a project admin, this choice
is lost so the user can see the updated project banner.

To test, you can set a project banner via the REST API with the REST
endpoint PUT /projects/:project_id/banner and hide the banner for a
specific user by setting a user pref with the name project_banner_<project_id>
to the value "hidden".

Part of story #12264: display a custom banner inside project pages

Change-Id: Iddd99af4eb716a63b8fbe5c74b6b28029cc80a25