Do not force deploy tuleap-svn-updater everywhere
tuleap-svn-updater was broken because it was running on systemd
based systems but was dealing with httpd services directly.

First thing was to fix tuleap-svn-updater to use the correct approach
to restart httpd depending on context (supervisord for distlp tests
and docker images like tuleap-aio:centos7 or systemd for regular
install).
In addition to that, there is actually no need to deploy this unit
on all servers as it's only needed in Distributed SVN setups. It
was decided to remove the deployment. A notice about the beahviour
change must be added to deployment guide to inform admins about
this manual thing on their side.

Fix request #13987 Cannot stop httpd due to tuleap worker

Change-Id: Iad47f848f5a218beb13647543073e884c3929ad9