Adjust apps height when a project banner is present
This is the same thing than for kanban (see 0b2c0e6468dfba91fa64f3b81349e6db9cc84882)
but for the taskboard and document apps. That seems to be the only
apps doing some calculations with the height of the top elements.

Part of story #12264: display a custom banner inside project pages

Change-Id: I227e4dd906f00bb60287ee339db69f4192b80721