App height takes into account the project banner
Unitl now, TestManagement was always loading the non condensed
stylesheet even when the user asked to use the condensed stylesheet.
Since the project banner have a different height if we are using the
condensed theme or not, this behavior has been fixed. The default
"PluginWithLegacyInternalRouting" has been replaced by an equivalent
custom implementation to allow us to load the right assets.

Part of story #12264: display a custom banner inside project pages

Change-Id: I001e7db5d91e4d136f1845e034456b7c31cc7f0d