Do not start old DB by default
People should have moved to MySQL 5.7 if they didn't have, the
will need to manually start the old db and do the migration.

The MySQL 5.7 service is now simply named "db" while the old
one has explicitly the version of MySQL in the name. This is
easier to manage across all the images because it can be
assumed the DB closer to the production environnements is
available at the "db" hostname. Developpers doing some
archeology or trying new version od the DB can manually do
the switch.

Part of request #13879 Switch development stack to mysql 5.7

Change-Id: I67d10a1c6f6b198ebaebf7076d86bb217f9980a7