Move all artifact creation/update transactions to DBTransactionExecutorWithConnection
The main purpose is to introduce a transaction at artifact creation for the creation
of first artifact changeset.

As there are already transactions in a few places (like Attachment field) and there is
a trick related to artifact link creation and hierarchy (a changeset is created automagically
behind the scene) we need to convert a bunch of transactions.

On a platform with ~200k artifacts the SQL gains are noticable: -31% of SQL time.

When we combine both (index on artifact_priority_rank), the creation of a
new artifact goes from 330ms to 173ms of wall time (-47%).
The SQL part goes from 188ms to 51ms (-73%).

Part of request #14048 Slowness at artifact creation

Change-Id: I1da6067e68c81ab114b4f7bf93fdbcda8dc4cad6