request #14518 Add cypress for dashboards
As of today widgets are not covered by e2e tests.
It's a specific entry point and it would be great to see durng nightly if we break something.
This patch add tests over some legacy widgets and gives us a coverage of cross tracker search usage.

Timetracking widgets should be covered, but I'll do it in a dediacted patch.

Change-Id: I59ed52ddcf2624817c54c38d0dedea471f8d3e8b