Improve create_test_env efficiency
Using mustache to parse a 40k lines XML file is not really efficient
It takes ~10s to parse the file on a prod instance. The usage of a
simple str_replace reduce to an hundred of ms.

Index on the artifact_priority table improve artifact creation timing
by 50% on platform with ~200k artifacts.

Part of request #14048 Slowness at artifact creation

Change-Id: I9f7931886817d40bdc85d7830a79422dcab461ac