Use route Rest to get burndown in release Widget and show errors
This is part of story #13300 See my progession with the burndown chart

Now, the plugin release widget uses route rest to get BurndownData of each releases.
The burndown chart is not yet created, but if there is an error to create it (no start date, under calculation, ...), it shows in the release.

How to test:
	- Install plugin release widget
	- Have some release with different burndown and errors (no start date, no duration, under calculation)

Expected result:
	- When you open a release:
		- If the burndown can be create, the message "There is nothing here!" is displayed
		- If there is an error, it is displayed

Change-Id: I7487404d2d54352443d56687af8cf07419dde9ff