Store expand/collapse swimlane in user prefs
Part of story #13799: Expand / collapse columns

In order to be consistent with the behavior of expand/collapse columns
(stored in user preferences), we need to change the expand/collapse of
swimlanes. There isn't anymore automagic collapse when the card is
closed. Instead we collapse only when the user decides to.

In addition, now we always load children of parent cards (even if it is
collapsed) because we will display the total number of cards in a column.
It is not ecofriendly, but since we suppose that there won't be that
much collapsed swimlane on a given taskboard, it is best to not clutter
much the data retrieval algorithm. Sorry, rainforest 🙀

Change-Id: I5f694fc30ba6e813cfb4072af3a03895356babde