Do not display A.D homepage in BurningParrot
Since git #tuleap/stable/29c89688e8b6b476a0ba638081439aa55fc99f60 the admin
page of A.D is displayed in BurningParrot (part of the admin page to be
more precise).

If a user tries to reach this page whereas she is not allowed to, then
we should not display the homepage in the current context, because
homepage is not BP compatible yet.

The quick fix is to redirect the user to the homepage, the behavior is
then a little bit better than before (the displayed content matches the
url of the page). A nicer approach would be to use front router and
throw a 403, it can be done later.

Part of request #13812: Migrate AgileDashboard administration to BurningParrot

Change-Id: I754426ba1acb8be992a853f60b085d9f7ef70422