Expand/Collapse column
User can now expand/collapse column.

Columns are expanded by default.

Given a column is collapsed,
Then every cards in it are hidden.

Given a solo card is in a collpased column,
Then the swimlane is not shown at all.

Given a column is collapsed,
Then collapsed state is stored in user prefs,
When user refresh the browser,
Then the column is still collapsed.

Given a column is expanded/collapsed for a milestone,
Then it is not expanded/collapsed for another milestone.

Given a column is expanded,
Then it can be expanded again.

Note: Cards count and vertical label will be done in a dedicated commit.
Note: Anonymous users prefs issue will be handled in a dedicated commit.

Part of story #13799: Expand / collapse columns

Change-Id: If4d0293640494dc8dfa50f619c1a36aa0b25a39a