Remove entirely the old "confirm_download" stuff
That was still use to force people to read the license agreement when
the file was referenced with a 'file' keyword.

The fix is not iso-func as it would have needed to deal with the modal
in Flaming Parrot and Burning Parrot on each and every page of the
app. That's not worth it for this kind of feature so, for the ones who
are relying on that, they are redirected on the parent release of the
file (in FRS) so they can get the popup confirmation.

The DB part is only purged for new references so developers can switch
on an old version of Tuleap and still have a working version of 'file'
keyword (so no forgeupgrade on purpose).

Part of story #14008 customize FRS Download Agreement in my project

Change-Id: If0cd65122c0583b9fc5c8e982a5580d86e7a677d